requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
idna==3.3
python-dotenv==0.19.1
requests==2.26.0
tqdm==4.62.3
urllib3==1.26.7
